Split `handler.py` file
